Multiple bug fix

+++ _editor/containers/EditorApp.jsx
@@ -30,6 +30,7 @@ import RichMarksModal from '../components/rich_plugins/rich_marks_modal/RichMark
 import AutoSave from '../components/autosave/AutoSave';
 import Clipboard from '../components/clipboard/Clipboard';
 import Alert from '../components/common/alert/Alert';
+import ToggleSwitch from '@trendmicro/react-toggle-switch';
 import i18n from 'i18next';
 import Ediphy from '../../core/editor/main';
 import { isSortableBox, isSection, isContainedView, isSortableContainer } from '../../common/utils';
@@ -147,16 +148,6 @@ class EditorApp extends Component {
                             });
 
                             this.dispatchAndSetState(deleteContainedView([cvid], boxesRemoving, containedViews[cvid].parent));
-
-                            Object.keys(containedViews[cvid].parent).forEach((el)=>{
-                                if (toolbars[el] && toolbars[el].state && toolbars[el].state.__marks) {
-                                    Ediphy.Plugins.get(toolbars[el].config.name).forceUpdate(
-                                        toolbars[el].state,
-                                        el,
-                                        DELETE_CONTAINED_VIEW
-                                    );
-                                }
-                            });
                         }}
                         onNavItemNameChanged={(id, titleStr) => this.dispatchAndSetState(changeNavItemName(id, titleStr))}
                         onNavItemAdded={(id, name, parent, type, position) => this.dispatchAndSetState(addNavItem(id, name, parent, type, position, (type !== 'section' || (type === 'section' && Ediphy.Config.sections_have_content))))}
@@ -445,7 +436,6 @@ class EditorApp extends Component {
                         let cvid = state.__marks[id].connection;
 
                         delete state.__marks[id];
-                        this.dispatchAndSetState(deleteRichMark(id, boxSelected, cvid, state));
                         // This checks if the deleted mark leaves an orphan contained view, and displays a message asking if the user would like to delete it as well
                         if (isContainedView(cvid)) {
                             let thiscv = containedViews[cvid];
@@ -468,21 +458,27 @@ class EditorApp extends Component {
                                         show
                                         hasHeader
                                         title={<span><i style={{ fontSize: '14px', marginRight: '5px' }} className="material-icons">delete</i>{i18n.t("messages.confirm_delete_cv")}</span>}
-                                        acceptButtonText={i18n.t("messages.confirm_delete_cv_as_well")}
-                                        cancelButton
-                                        cancelButtonText={i18n.t("messages.confirm_delete_cv_not")}
-                                        closeButton onClose={(bool)=>{
-                                            if (bool) {
-                                                let boxesRemoving = [];
-                                                containedViews[cvid].boxes.map(boxId => {
-                                                    boxesRemoving.push(boxId);
-                                                    boxesRemoving = boxesRemoving.concat(this.getDescendantBoxes(boxes[boxId]));
-                                                });
+                                        acceptButtonText={i18n.t("messages.OK")}
 
-                                                this.dispatchAndSetState(deleteContainedView([cvid], boxesRemoving, thiscv.parent));
+                                        closeButton onClose={(bool)=>{
+
+                                            if (bool) {
+                                                this.dispatchAndSetState(deleteRichMark(id, boxSelected, cvid, state));
+                                                let deleteAlsoCV = document.getElementById('deleteAlsoCv').classList.toString().indexOf('checked') > 0;
+                                                if(deleteAlsoCV) {
+                                                    let boxesRemoving = [];
+                                                    containedViews[cvid].boxes.map(boxId => {
+                                                        boxesRemoving.push(boxId);
+                                                        boxesRemoving = boxesRemoving.concat(this.getDescendantBoxes(boxes[boxId]));
+                                                    });
+
+                                                    this.dispatchAndSetState(deleteContainedView([cvid], boxesRemoving, thiscv.parent));
+                                                }
                                             }
                                             this.setState({ alert: null });}}>
-                                        <span> {confirmText} </span>
+                                        <span> {confirmText} </span><br/>
+                                        <ToggleSwitch id="deleteAlsoCv" />
+                                        {i18n.t("messages.confirm_delete_cv_as_well")}
                                     </Alert>);
                                     this.setState({ alert: alertComponent });
                                 }
